Fix the provider config tests too
